% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardlistr_functions.R
\name{check_dir_exists}
\alias{check_dir_exists}
\title{Check if file exists}
\usage{
check_dir_exists(dir)
}
\arguments{
\item{dir}{Path to directory}
}
\description{
Check if file exists
}
\keyword{internal}
