% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardlistr_functions.R
\name{clean_ebird_data}
\alias{clean_ebird_data}
\title{Extract eBird data for specified location}
\usage{
clean_ebird_data(file_data, location, call = rlang::caller_env())
}
\arguments{
\item{file_data}{Path to csv file containing eBird data}

\item{location}{A string containing the exact name of the location
to extract from the data set}

\item{call}{The environment from which the function is called}
}
\value{
Tibble in tidy data format
}
\description{
Extract eBird data for specified location
}
