% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardlistr_functions.R
\name{get_time}
\alias{get_time}
\title{Extract time from lubridate datetime}
\usage{
get_time(time = lubridate::now())
}
\arguments{
\item{time}{A vector of date-time values}
}
\value{
A vector of times stored as number of seconds since 00:00:00. See \code{hms::hms()}
}
\description{
Extract time from lubridate datetime
}
\keyword{internal}
