% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardlistr_functions.R
\name{select_file}
\alias{select_file}
\title{Make best guess which file from input folder to use}
\usage{
select_file(dir_dat)
}
\arguments{
\item{dir_dat}{Path to an existing directory containing eBird data}
}
\value{
The last csv file in ascending alphanumeric order
}
\description{
Make best guess which file from input folder to use
}
\keyword{internal}
