% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardlistr_functions.R
\name{tetrad}
\alias{tetrad}
\title{Calculate tetrad from a datetime}
\usage{
tetrad(x, method = "ebird")
}
\arguments{
\item{x}{Vector of datetimes.}

\item{method}{Method to calculate tetrad.
\itemize{
\item \code{ebird} (the default): The "ebird" method takes day 1-7 as tetrad 1, 8-14 as tetrad 2, 15-21 as tetrad 3 and >21 as tetrad 4. This method is employed by the Cornell Lab of Ornithology in their bar charts on the eBird platform
\item \code{fourths}: the "fourths" method divides a month into four tetrads that are as close to equal in length as possible. A month with 30 days will consists of tetrads of length 7, 8, 7 and 8 in this order. A month with 28 days will consist of tetrads of length 7
}}
}
\value{
Vector of tetrad assignments
}
\description{
A tetrad is one of four parts of a month. Since most months are not
a multiple of a week long (except for February in non-leap years),
tetrads do not have the same length in days.
}
