% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardlistr_functions.R
\name{unique_values}
\alias{unique_values}
\title{Return list of unique values without changing their type}
\usage{
unique_values(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A sorted vector of unique values
}
\description{
Return list of unique values without changing their type
}
\keyword{internal}
