% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardlistr-package.R
\docType{package}
\name{yardlistr-package}
\alias{yardlistr-package}
\title{yardlistr: Generate Summary Plots from eBird Data}
\description{
The yardlistr package loads your personal eBird data from a csv file, processes checklists from a user-specified list of locations and generates a variety of plots for each location.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gl-eb/yardlistr}
  \item \url{https://www.gl-eb.me/yardlistr/}
  \item Report bugs at \url{https://github.com/gl-eb/yardlistr/issues}
}

}
\author{
\strong{Maintainer}: Gleb G. Ebert \email{dev@gl-eb.me} (\href{https://orcid.org/0000-0002-7734-3824}{ORCID}) [copyright holder]

}
\keyword{internal}
