% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardlistr_main.R
\name{yardlistr}
\alias{yardlistr}
\title{Main yardlistr functions that calls all other helper functions}
\usage{
yardlistr(location, dir_dat, dir_img)
}
\arguments{
\item{location}{(character) Name of location as it appears in eBird data}

\item{dir_dat}{(character) Relative or absolute path to directory containing
eBird data}

\item{dir_img}{(character) Relative or absolute path to directory to which
plots should be saved}
}
\value{
A list containing the tibbles used to generate plots
}
\description{
Main yardlistr functions that calls all other helper functions
}
\examples{
\dontrun{
  yardlistr("My Location", "./Data", "./Plots")
}
}
